Getting rid of old redendencies caused a bit of restructurising of gem. It is now not dependent on debugger
