+++ .gitignore
@@ -46,3 +46,5 @@ pkg
 
 # For rubinius:
 #*.rbc
+
+nbproject/*
