+++ .rvmrc
@@ -1,0 +1,1 @@
+rvm --create use 1.9.3@happy_fun_time_bot
