+++ Gemfile
@@ -1,13 +1,18 @@
 source "http://rubygems.org"
-gem 'ruby-debug'
+
 gem 'xmpp4r'
 gem 'json'
 gem 'httparty'
 
 group :development do
-  gem "rspec", "~> 2.3.0"
-  gem "yard", "~> 0.6.0"
-  gem "bundler", "~> 1.0.0"
-  gem "jeweler", "~> 1.6.0"
-  gem "rcov", ">= 0"
+  gem "rspec", "~> 2.10.0"
+  gem "jeweler", "~> 1.8.0"
+  gem "yard", "~> 0.8.1"
+  if RUBY_VERSION >= '1.9.0'
+    gem "debugger", "~> 1.1.3"
+    gem "simplecov", "~> 0.6.4"
+  else
+    gem 'ruby-debug'
+    gem "rcov", "~> 1.0.0"
+  end
 end
